requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
Werkzeug==3.0.1
PyJWT==2.8.0

